Move Slack reconciler code to pkg/reconciler and cmd/
