+++ config/200-clusterroles.yaml
@@ -50,7 +50,8 @@ rules:
 # Read Source resources and update their statuses
 - apiGroups:
   - sources.triggermesh.io
-  resources: []
+  resources:
+  - slacksources
   verbs:
   - list
   - watch
@@ -58,7 +59,8 @@ rules:
   - patch
 - apiGroups:
   - sources.triggermesh.io
-  resources: []
+  resources:
+  - slacksources/status
   verbs:
   - update
 
@@ -116,7 +118,8 @@ metadata:
 rules:
 - apiGroups:
   - sources.triggermesh.io
-  resources: []
+  resources:
+  - slacksources
   verbs:
   - get
   - list
