+++ config/300-slacksource.yaml
