+++ pkg/reconciler/slacksource/adapter.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package reconciler
+package slacksource
 
 import (
 	corev1 "k8s.io/api/core/v1"
