--- slack/Makefile
@@ -1,9 +1,0 @@
-KSOURCE      = slack
-KSOURCE_DESC = Triggermesh Slack Event Source for Knative
-BASE_DIR    := $(CURDIR)/..
-
-include $(BASE_DIR)/scripts/inc.Makefile
-
-GOPKGS = ./cmd/... ./pkg/reconciler/...
-
-include $(BASE_DIR)/scripts/inc.Target
