--- slack/cloudbuild.yaml
@@ -1,28 +1,0 @@
-steps:
-
-- name: 'gcr.io/kaniko-project/executor:latest'
-  args:
-  - --dockerfile=${_KSOURCE}/cmd/controller/Dockerfile
-  - --destination=gcr.io/$PROJECT_ID/${_KSOURCE}-source-controller:${COMMIT_SHA}
-  - --destination=gcr.io/$PROJECT_ID/${_KSOURCE}-source-controller:${_KANIKO_IMAGE_TAG}
-  - --cache-repo=gcr.io/$PROJECT_ID/${_KSOURCE}-source-controller/cache
-  - --cache=${_KANIKO_USE_BUILD_CACHE}
-  - --no-push=${_KANIKO_NO_PUSH}
-  - ${_KANIKO_EXTRA_ARGS}
-  waitFor: ['-']
-
-timeout: 1800s
-
-substitutions:
-  _KSOURCE: "slack"
-  _KANIKO_IMAGE_TAG: "latest"
-  _KANIKO_NO_PUSH: "false"
-  _KANIKO_USE_BUILD_CACHE: "true"
-  _KANIKO_EXTRA_ARGS: ""
-
-options:
-  substitution_option: 'ALLOW_LOOSE'
-
-tags:
-  - slack
-  - knative-sources
