--- slack/cmd/controller/kodata/HEAD
@@ -1,1 +1,0 @@
-../../../../.git/HEAD
