--- slack/cmd/controller/kodata/refs
@@ -1,1 +1,0 @@
-../../../../.git/refs
