--- slack/config/500-controller.yaml
@@ -1,68 +1,0 @@
-# Copyright (c) 2020 TriggerMesh Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: &app slack-source-controller
-  namespace: triggermesh
-  labels:
-    control-plane: slack-source-controller-manager
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      control-plane: *app
-
-  template:
-    metadata:
-      labels:
-        control-plane: *app
-
-    spec:
-      serviceAccountName: *app
-
-      containers:
-      - name: controller
-        terminationMessagePolicy: FallbackToLogsOnError
-        image: ko://github.com/triggermesh/knative-sources/slack/cmd/controller
-        resources:
-          requests:
-            cpu: 20m
-            memory: 20Mi
-
-        env:
-          - name: SYSTEM_NAMESPACE
-            valueFrom:
-              fieldRef:
-                fieldPath: metadata.namespace
-          # Logging/observability configuration
-          - name: CONFIG_LOGGING_NAME
-            value: config-logging
-          - name: CONFIG_OBSERVABILITY_NAME
-            value: config-observability
-          - name: METRICS_DOMAIN
-            value: triggermesh.io/sources
-          # Adapters
-          - name: SLACKSOURCE_ADAPTER_IMAGE
-            value: ko://github.com/triggermesh/knative-sources/slack/cmd/adapter
-
-        securityContext:
-          allowPrivilegeEscalation: false
-
-        ports:
-        - name: metrics
-          containerPort: 9090
-        - name: profiling
-          containerPort: 8008
